dbgraph: convert numSamples to int
The getNumSamples() accessor returns `int`. `numSamples` also compares
against `xdata.size()` which returns `int`.
Makes sense to convert `numSamples` to `int`. If we need more than 2
billion samples, we can bump this to 64 bit int.

+++ src/dbgraph.cpp
@@ -262,7 +262,7 @@ const QwtScaleWidget *dBgraph::getAxisWidget(QwtAxisId id)
 
 void dBgraph::setNumSamples(int num)
 {
-	numSamples = (unsigned int) num;
+	numSamples = num;
 
 	reset();
 	ydata.reserve(numSamples + 1);
